Build cardscan example app
