+++ build.gradle
@@ -22,6 +22,7 @@ buildscript {
 plugins {
     id 'io.github.gradle-nexus.publish-plugin' version '1.1.0'
     id 'io.codearte.nexus-staging' version '0.30.0'
+    id 'org.jetbrains.kotlin.android' version '1.5.31' apply false
 }
 
 allprojects {
