+++ stripecardscan-example/build.gradle
@@ -50,6 +50,8 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-viewmodel-compose:2.4.0"
     // Integration with observables
     implementation "androidx.compose.runtime:runtime-livedata:$composeVersion"
+    implementation 'androidx.appcompat:appcompat:1.3.0'
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
 
     debugImplementation 'com.squareup.leakcanary:leakcanary-android:2.8.1'
