+++ stripecardscan-example/src/main/AndroidManifest.xml
@@ -21,14 +21,22 @@
 
         <activity
             android:name=".MainActivity"
-            android:label="@string/app_name"
-            android:theme="@style/AppTheme.NoActionBar"
-            android:exported="true">
+            android:exported="true"
+            android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity
+            android:name=".CardImageVerificationDemoActivity"
+            android:exported="false" />
+
+        <activity
+            android:name=".CardScanDemoActivity"
+            android:exported="false" />
     </application>
 
 </manifest>
