+++ stripecardscan-example/src/main/res/values/strings.xml
@@ -1,6 +1,7 @@
 <resources xmlns:tools="http://schemas.android.com/tools" tools:ignore="MissingTranslation">
     <string name="app_name">StripeCardScan Example</string>
     <string name="cardimageverificationsheet">CardImageVerificationSheet</string>
+    <string name="cardscansheet">CardScanSheet</string>
 
     <string name="getPublishableKey">Uset Test Publishable Key</string>
     <string name="stripePublishableKey">Stripe Publishable Key</string>
